Hn1

+++ include/domain/entities/Staff.hpp
@@ -4,11 +4,11 @@
 
 class Staff : public Account {
 private:
-    float salary;
+    long salary;
 
 public:
     Staff() : Account(), salary(0.0) {}
-    Staff(const string &Id, const string &Name, const string &Pass, const string &Gen, const float &Salary) : Account(Id, Name, Pass, Gen), salary(Salary) {}
-    float getSalary() const { return salary; };              // getter
-    void setSalary(float salary) { this->salary = salary; }; // setter
+    Staff(const string &Id, const string &Name, const string &Pass, const string &Gen, const long &Salary) : Account(Id, Name, Pass, Gen), salary(Salary) {}
+    long getSalary() const { return salary; };              // getter
+    void setSalary(long salary) { this->salary = salary; }; // setter
 };
